Island: update story & split off IslandGroup story
